Create stage module and move the JoiningPeer stages there
Also rename `BootstrappingStage` to `Bootstrapping` and `JoiningStage` to `Joining`.
